调试通过购进新增、提交功能。

+++ mf-item/mf-item-interface/src/main/java/com/mallfe/item/pojo/Kucn.java
@@ -32,5 +32,5 @@ public class Kucn {
     /**
      * 库存
      */
-    private Long kucn;
+    private Integer kucn;
 }
