+++ mf-item/mf-item-interface/src/main/java/com/mallfe/item/pojo/KucnOut.java
@@ -29,7 +29,7 @@ public class KucnOut{
      * 类型
      */
     private Integer lx;
-    private Long sl;
+    private Integer sl;
     private String lsh;
     private Integer xh;
     private String ywbm;
